+++ benito-frontend/src/actions/home/index.tsx
@@ -1,8 +1,7 @@
-import { Project, Category } from "../../types";
+import { Project } from "../../types";
 import {
   HomeAction,
   UPDATE_FEATURED_PROJECTS,
-  UPDATE_CATEGORIES,
   UPDATE_LATEST_PROJECTS,
   UPDATE_PROJECT_TOTAL,
 } from "../../store/home/types";
@@ -18,13 +17,6 @@ export function updateLatestProjects(projects: Array<Project>): HomeAction {
   return {
     type: UPDATE_LATEST_PROJECTS,
     payload: projects,
-  };
-}
-
-export function updateCategories(categories: Array<Category>): HomeAction {
-  return {
-    type: UPDATE_CATEGORIES,
-    payload: categories,
   };
 }
