+++ benito-frontend/src/reducers/index.ts
@@ -3,12 +3,14 @@ import { combineReducers } from "redux";
 import userReducer from "./user";
 import homeReducer from "./home";
 import searchReducer from "./search";
+import commonReducer from "./common";
 
 const rootReducer = combineReducers({
   login: loginReducer,
   user: userReducer,
   search: searchReducer,
   home: homeReducer,
+  common: commonReducer,
 });
 
 export type RootState = ReturnType<typeof rootReducer>;
