+++ benito-frontend/src/store/home/types.tsx
@@ -1,8 +1,7 @@
-import { Project, Category } from "../../types";
+import { Project } from "../../types";
 
 export const UPDATE_FEATURED_PROJECTS = "UPDATE_FEATURED_PROJECTS";
 export const UPDATE_LATEST_PROJECTS = "UPDATE_LATEST_PROJECTS";
-export const UPDATE_CATEGORIES = "UPDATE_CATEGORIES";
 export const UPDATE_PROJECT_TOTAL = "UPDATE_PROJECT_TOTAL";
 
 interface UpdateFeaturedProjectsAction {
@@ -15,11 +14,6 @@ interface UpdateLatestProjectsAction {
   payload: Array<Project>;
 }
 
-interface UpdateCategoriesAction {
-  type: typeof UPDATE_CATEGORIES;
-  payload: Array<Category>;
-}
-
 interface UpdateProjectTotalAction {
   type: typeof UPDATE_PROJECT_TOTAL;
   payload: Number;
@@ -27,13 +21,11 @@ interface UpdateProjectTotalAction {
 
 export type HomeAction =
   | UpdateFeaturedProjectsAction
-  | UpdateCategoriesAction
   | UpdateLatestProjectsAction
   | UpdateProjectTotalAction;
 
 export type HomeState = {
   featuredProjects: Array<Project>;
   latestProjects: Array<Project>;
-  categories: Array<Category>;
   projectTotal: Number;
 };
